Fix a use-after-free error.
Capturing a function from a `sol::state` keeps a pointer to the
associated `sol::state`.  If this is moved, the functions point to freed
memory.  When the `sol::state` objects were stored in a `std::vector`,
resizing the vector caused UAF.

This is fixed by storing a shared pointer to the LuaMatcher in the
vector instead.  It's not worth caring about the overhead of a couple of
extra words of memory and one dereference for an object that
encapsulates a complete Lua VM.